ajuste na logica de registro de usuário para quando não existir nenhum usuário cadastrado e ajustes nas entidades com propriedades decimal passando para float

+++ Bag-and-Shop-app/Domain/Entities/Bag.cs
@@ -5,7 +5,7 @@ namespace Bag_and_Shop_app.Domain.Entities
     public class Bag
     {
         public int Id { get; set; }
-        public decimal Gold { get; set; } = 0;
+        public int Gold { get; set; } = 0;
         public double WeightLimit { get; set; } = 20.0;
         public double CurrentWeight { get; set; } = 0.0;
         public int CharacterId { get; set; }
